src/main/java/com/fastcampus/javaallinone/project3/demo/controller/HelloWorldController.java
package com.fastcampus.javaallinone.project3.demo.controller;

import com.fastcampus.javaallinone.project3.demo.Exception.dto.ErrorResponse;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RestController;

@Slf4j
@RestController
public class HelloWorldController {

    @PostMapping("api/helloWorld")
    public String helloWorld(){
        return "hello world";
    }

    @GetMapping("/api/helloException")
    public String helloException(){
        throw new RuntimeException("Hello RuntimeException");
    }
}
